[webview_flutter] Migrate main package to fully federated architecture. (#4366)
