--- packages/webview_flutter/webview_flutter/android/build.gradle
@@ -1,57 +1,0 @@
-group 'io.flutter.plugins.webviewflutter'
-version '1.0-SNAPSHOT'
-
-buildscript {
-    repositories {
-        google()
-        mavenCentral()
-    }
-
-    dependencies {
-        classpath 'com.android.tools.build:gradle:3.3.0'
-    }
-}
-
-rootProject.allprojects {
-    repositories {
-        google()
-        mavenCentral()
-    }
-}
-
-apply plugin: 'com.android.library'
-
-android {
-    compileSdkVersion 29
-
-    defaultConfig {
-        minSdkVersion 19
-        testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
-    }
-
-    lintOptions {
-        disable 'InvalidPackage'
-        disable 'GradleDependency'
-    }
-
-    dependencies {
-        implementation 'androidx.annotation:annotation:1.0.0'
-        implementation 'androidx.webkit:webkit:1.0.0'
-        testImplementation 'junit:junit:4.12'
-        testImplementation 'org.mockito:mockito-inline:3.11.1'
-        testImplementation 'androidx.test:core:1.3.0'
-    }
-
-
-    testOptions {
-        unitTests.includeAndroidResources = true
-        unitTests.returnDefaultValues = true
-        unitTests.all {
-            testLogging {
-               events "passed", "skipped", "failed", "standardOut", "standardError"
-               outputs.upToDateWhen {false}
-               showStandardStreams = true
-            }
-        }
-    }
-}
