--- packages/webview_flutter/webview_flutter/android/settings.gradle
@@ -1,1 +1,0 @@
-rootProject.name = 'webview_flutter'
