--- packages/webview_flutter/webview_flutter/android/src/main/AndroidManifest.xml
@@ -1,2 +1,0 @@
-<manifest package="io.flutter.plugins.webviewflutter">
-</manifest>
