--- packages/webview_flutter/webview_flutter/ios/Classes/FLTWebViewFlutterPlugin.h
@@ -1,8 +1,0 @@
-// Copyright 2013 The Flutter Authors. All rights reserved.
-// Use of this source code is governed by a BSD-style license that can be
-// found in the LICENSE file.
-
-#import <Flutter/Flutter.h>
-
-@interface FLTWebViewFlutterPlugin : NSObject <FlutterPlugin>
-@end
